mugo-zap/server/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.6
gotrue==2.12.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.1.0
idna==3.11
jiter==0.13.0
openai==2.17.0
packaging==26.0
postgrest==0.16.11
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
websockets==12.0
